V05

+++ app/src/main/java/fr/jlt/gdpw/utils/ImportMiniature.java
@@ -10,6 +10,8 @@ import java.io.FileNotFoundException;
 import java.util.Scanner;
 
 import fr.jlt.gdpw.donneesDAO.Miniature;
+import fr.jlt.gdpw.execption.ModeleObligatoireException;
+import fr.jlt.gdpw.execption.RubriquesObligatoiresException;
 import fr.jlt.gdpw.table.MiniatureBDD;
 
 /**
@@ -55,7 +57,7 @@ public class ImportMiniature {
         //items = line.split("\\|");
         String[] items = line.split("\\|", -1);    //TO DO vérifier le -1 !
         if (!"Modele".equals(items[0].trim())) {
-            ajout = ajoutMiniature(bdd, line, path, externalFilesDir);
+            ajout = ajoutMiniature(bdd, line, path, externalFilesDir, 0);
         }
         return ajout;
     }
@@ -67,22 +69,44 @@ public class ImportMiniature {
      * Convertir une ligne du fichier en enregistrement dans la table
      * @return Table
      */
-    public static Boolean ajoutMiniature(final SQLiteDatabase bdd, final String line, final String path, final String externalFilesDir) {
+    public static Boolean ajoutMiniature(final SQLiteDatabase bdd, final String line, final String path, final String externalFilesDir, final int doublon) {
         Boolean ajout = null;
         String[] items = line.split("\\|", -1);
+        if (items.length < 3) {
+            throw new RubriquesObligatoiresException();
+        }
+
         Miniature miniature = new Miniature();
 
         int i = 0;
 
         miniature.setModele(items[i++]);
+        if (Utils.isEmptyOrNullOrBlank(miniature.getModele())) {
+            throw new ModeleObligatoireException();
+        }
 
         String trouve;
         try {
-            //Log.d("ImportMiniature", "mise à jour : " + miniature.getModele());
             Miniature miniatureOld = MiniatureBDD.get(miniature.getModele(), bdd);
             trouve = miniatureOld.getTrouve();
-            MiniatureBDD.delete(miniature.getModele(), bdd);
-            ajout = false;
+            switch(doublon) {
+                case 0:
+                    // remplacer la fiche existante
+                    //Log.d("ImportMiniature", "doublon remplacé : " + miniature.getModele());
+                    MiniatureBDD.delete(miniature.getModele(), bdd);
+                    ajout = false;
+                    break;
+                case 1:
+                    // garder la fiche existante
+                    //Log.d("ImportMiniature", "doublon conservé : " + miniature.getModele());
+                    ajout = null;
+                    break;
+                case 2:
+                    // creer un doublon
+                    //Log.d("ImportMiniature", "doublon créé : " + miniature.getModele());
+                    ajout = true;
+                    break;
+            }
         }
         catch (Exception e) {
             //Log.d("ImportMiniature", "ajout : " + miniature.getModele());
@@ -90,37 +114,50 @@ public class ImportMiniature {
             ajout = true;
         }
 
-        miniature.setMarque(items[i++]);
-        miniature.setCollection(items[i++]);
-        miniature.setPreference(items[i++]);
-        miniature.setReference(items[i++]);
-        miniature.setPrix(items[i++]);
-        miniature.setDateSortie(items[i++].replaceAll("-", "/"));
-        miniature.setCarrosserie(items[i++]);
-        //miniature.setPhoto(ctx.getResources().getIdentifier(items[i++], "drawable", ctx.getPackageName()));
-        miniature.setPhoto(items[i++]);
-        if (!"/".equals(miniature.getPhoto().substring(0, 1)))  {
-            miniature.setPhoto("/" + miniature.getPhoto());
-        }
-        miniature.setEditeur(items[i++]);
-        miniature.setFabricant(items[i++]);
-        miniature.setTrouve(trouve);
+        if (ajout != null) {
+            try {
+                miniature.setMarque(items[i++]);
+                miniature.setPreference(items[i++]);
+                //miniature.setPhoto(ctx.getResources().getIdentifier(items[i++], "drawable", ctx.getPackageName()));
+                miniature.setPhoto(items[i++]);
+                if (!"/".equals(miniature.getPhoto().substring(0, 1))) {
+                    miniature.setPhoto("/" + miniature.getPhoto());
+                }
+                miniature.setCollection(items[i++]);
+                miniature.setReference(items[i++]);
+                miniature.setPrix(items[i++]);
+                miniature.setDateSortie(items[i++].replaceAll("-", "/"));
+                miniature.setCarrosserie(items[i++]);
+                miniature.setEditeur(items[i++]);
+                miniature.setFabricant(items[i++]);
 //Log.d("ImportMiniature", "Mini : " + miniature.toString());
+            }
+            catch (IndexOutOfBoundsException e) {
+                // ne rien faire
+            }
 
-        MiniatureBDD.add(miniature, bdd);
+            if (miniature.getPreference() == null) {
+                miniature.setPreference("0");
+            }
+            miniature.setTrouve(trouve);
 
-        //Déplace la photo dans l'espace privée de l'appli.
-        File fileSrc = new File(path + miniature.getPhoto());
-        File fileDest = new File(externalFilesDir, miniature.getPhoto());
-        Utils.makeDirs(fileDest);
-        Utils.copyFile(fileSrc, fileDest);
-        fileSrc.delete();
+            MiniatureBDD.add(miniature, bdd);
 
-        //crée une miniature et la met en cache
-        File fileCacheDest = new File(externalFilesDir + "/cached", miniature.getPhoto());
-        Utils.makeDirs(fileCacheDest);
-        Bitmap bmp = Utils.decodeSampledBitmapFromUri(fileDest.getAbsolutePath(), 200, 150);
-        Utils.StoreImage(bmp, fileCacheDest);
+            if (Utils.isNotEmptyOrNull(miniature.getPhoto())) {
+                //Déplace la photo dans l'espace privée de l'appli.
+                File fileSrc = new File(path + miniature.getPhoto());
+                File fileDest = new File(externalFilesDir, miniature.getPhoto());
+                Utils.makeDirs(fileDest, true);
+                Utils.copyFile(fileSrc, fileDest);
+                fileSrc.delete();
+
+                //crée une miniature et la met en cache
+                File fileCacheDest = new File(externalFilesDir + "/cached", miniature.getPhoto());
+                Utils.makeDirs(fileCacheDest, true);
+                Bitmap bmp = Utils.decodeSampledBitmapFromUri(fileDest.getAbsolutePath(), 200, 150);
+                Utils.StoreImage(bmp, fileCacheDest);
+            }
+        }
 
         return ajout;
     }
